Dump keys after 60 seconds of inactivity
Currently we dump the keypresses after 60 seconds. This often leads to phrases being split across protobuf boundaries. While the protobufs can be merged, it makes more sense to dump in between utterances. Since we don't know for sure, waiting for a long enough period of inactivity (ie: no keypresses) is a reasonable choice.