[WebUI] Accept access_token route parameter (#170)
Fixes #164

This allows the host app, which *can* persist refresh tokens, to
pass the access token directly into the WebUI. This gets rid of the
need to do limited-input OAuth. 

The host C# app performs the usual OAuth workflow as shown in the example:
https://github.com/googlesamples/oauth-apps-for-windows

and stores the refresh token, which can be used to acquire new access tokens
in the future (after app or system restart).