[WebUI] Add QuickPhrasesComponent, MiniBarComponent, and related app states (#171)
Fixes #165.

The `QuickPhrasesComponent` is composed of `PhraseComponent`s that are controlled by gaze-clickable scroll-up and scroll-down buttons. See screenshot: 
![image](https://user-images.githubusercontent.com/16824702/151589493-f224a915-1c1d-4274-91c1-855d196b4ee9.png)

Currently, the phrases input to `QuickPhrasesComponent` instances are hardcoded in the `AppComponent`. Follow-up PRs will retrieve them from a remote API endpoint. 

The `AppComponent` now has a number of AppState enum values controlling its overall state. See the left side of the screenshot above for the left nav buttons that allows the user to switch between the different quick phrase sets and the abbreviation expansion mode. The button at the bottom-left corner minimizes the app to a gaze-clickable button that can expand the app back to the previous non-minimized state. See screenshot:
![image](https://user-images.githubusercontent.com/16824702/151589786-a3006257-726a-4531-b278-76cfcb0ef394.png)

These buttons are all gaze-clickable.

Related changes in this PR:
- Unit tests are added for the new components: `QuickPhrasesComponent` and `MiniBarCompoent`
- Unit tests are updated for the existing component: `AppComponent`
- In `ExternalEventsComponent`, allow unregistering of previously registered keypress callbacks. This is now necessary because `AbbreviationComponent` can now get destroyed and re-created in the `AppComponent` depending on its state. So we want to avoid leaking previous callback registrations.
- In cefsharp.ts API, allow specification of a container (parent) view port (DOMRect), so that elements that are invisible due to scroll state can be skipped when registering gaze-clickable button boxes. This is used in `QuickPhrasesComponent` when there are many elements causing overflow and scrolling. See screenshot above.