[WebUI] Add settings.ts, version.ts and SettingsComponent (#205)
Add a settings page. It allows the user to change settings such as TTS voice type and TTS volume. It also contains a 
"Reload" button that allows the user to force-reload the app. See screenshot below:
![image](https://user-images.githubusercontent.com/16824702/154851628-773c0c2c-048b-4ce2-b9a8-31936f5aae09.png)

The settings are stored in one of two possible places:
- Via cefsharp.ts saveSettings() and loadSettings() API, which will request the native layer to save/load settings.
  See the updated README.md file
- If cefsharp's bound listener object is unavailable, in browser local storage.

Towards #198: The settings logic is not connected with TextToSpeechComponent yet. The branching of the logic for
personalized (cloud) voice and generic (local) voice will be done in another PR.
Fixes #201: The version string (e.g., v0.0.1) is displayed at the top of the settings page.

Also in this PR: Rearrange the icons in the left nav:
- Move the "Minimize" button to the top. Use the "close" icon
- Put the settings (menu) icon at the bottom.