[WebUI] Support basic cursor navigation in InputBarComponent (#311)
Click events in the `InputBarComponent`'s input-text div previous didn't get handled. In this PR, the events are now handled and are used to place the cursor in the corresponding location.

This PR also fixes the logical for word-backspace (Ctrl+Backspace or Ctrl,Shift,Left,Backspace) for non-final cursor locations.

When cursor is not at the final location, it can insert text into the text in the input bar.

See screenshot:
![image](https://user-images.githubusercontent.com/16824702/167274497-9b01887d-2464-45e7-877a-e1ef261500e2.png)

Known issues: When there is line wrapping in input-text, placing the cursor changes the wrapping, because a line break is inserted right after the cursor.